elevation_layer renamed as elevation_maping_costmap_2d_plugin. Somehow doesn't build
