--- elevation_layer/costmap_plugins.xml
@@ -1,7 +1,0 @@
-<class_libraries>
-  <library path="lib/libelevation_layer">
-    <class type="elevation_layer::ElevationLayer" base_class_type="costmap_2d::Layer">
-      <description>adds elevation map info to costmap_2d.</description>
-    </class>
-  </library>
-</class_libraries>
