+++ elevation_maping_costmap_2d_plugin/cfg/ElevationPlugin.cfg
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
-PACKAGE = "elevation_layer"
+PACKAGE = "elevation_mapping_costmap_2d_plugin"
 
 from dynamic_reconfigure.parameter_generator_catkin import ParameterGenerator, bool_t, double_t, int_t
 
 gen = ParameterGenerator()
 gen.add("enabled", bool_t, 0, "Whether to apply this plugin or not", True)
 
-exit(gen.generate("elevation_layer", "elevation_layer", "ElevationPlugin"))
+exit(gen.generate("elevation_mapping_costmap_2d_plugin", "elevation_mapping_costmap_2d_plugin", "ElevationPlugin"))
